New Screens
I added new screens to the project and fully created my second screen.
The new screens are blank and need to be made. Also I have made changes
to color and created a new dimention for text. Also I added in a new
color for the background of my buttons.